+++ apps/freelance/src/pages/ChatPage/index.tsx
@@ -2,15 +2,9 @@ import { Avatar, Form, Row } from 'antd';
 import { useTranslation } from 'react-i18next';
 import { UserOutlined } from '@ant-design/icons';
 import { InputSearch, StyledInput } from '@freelance/components';
-import useAppSelector from '@hooks/useAppSelector';
-import {
-  useGetMessagesQuery,
-  useSendMessageMutation,
-} from 'redux/services/chatApi';
-import { useGetUserInfoQuery } from 'redux/services/user';
 import { formatDate, formatTime } from 'src/utils/dates';
 
-import { conversation, users } from './constants';
+import { users } from './constants';
 import {
   BottomWrapper,
   FirstUserContainer,
@@ -25,36 +19,11 @@ import {
   StyledRightSide,
   UserWrapper,
 } from './styles';
-
-type InputType = {
-  message: string;
-};
-
-type MessageType = {
-  token: string;
-  conversation: number;
-  message: string;
-};
+import useChatData from './useChatData';
 
 const ChatPage = () => {
   const { t } = useTranslation();
-  const { access_token }: { access_token: string } = useAppSelector(
-    state => state.user,
-  );
-  const token = access_token;
-  const { data: dataUser } = useGetUserInfoQuery();
-  const { data: dataMessages } = useGetMessagesQuery({ token, conversation });
-  const [form] = Form.useForm<InputType>();
-  const [sendMessage] = useSendMessageMutation();
-
-  const handleSend = (values: InputType) => {
-    const message: MessageType = {
-      token: access_token,
-      conversation: conversation,
-      message: values.message,
-    };
-    message.token && sendMessage(message);
-  };
+  const { user, chatMessages, form, handleSend } = useChatData();
 
   return (
     <Row>
@@ -79,12 +48,12 @@ const ChatPage = () => {
           <h2>{t('chat.user')}</h2>
         </HeaderContainer>
         <MessagesWrapper>
-          {dataMessages &&
-            dataMessages.map(item =>
-              dataUser?.email === item.from.email ? (
+          {chatMessages &&
+            chatMessages.map(item =>
+              user?.email === item.from.email ? (
                 <FirstUserContainer>
                   <p>
-                    {dataUser.first_name} {dataUser.last_name}{' '}
+                    {user.first_name} {user.last_name}{' '}
                     {formatDate(new Date(item.created_at))}{' '}
                     {formatTime(new Date(item.created_at))}
                   </p>
